fix the use of raw session ids for things that rly needed to use makeSession

+++ src/modules/ui/hooks/GauntletLayer.cpp
@@ -118,7 +118,7 @@ struct GLOBED_MODIFY_ATTR HookedGauntletLayer : geode::Modify<HookedGauntletLaye
         if (!nm.isConnected()) return;
 
         auto levelIds = asp::iter::from(CCArrayExt<GJGameLevel*>(m_levels))
-            .map([](auto level) { return (uint64_t) level->m_levelID; })
+            .map([](auto level) { return (int) level->m_levelID; })
             .collect();
 
         nm.sendRequestPlayerCounts(std::span{levelIds.data(), levelIds.size()});
